Putting the nested class back in the same file as the parent.
